[IMPROVE] way to define props to send
