+++ example/views/SearchOptionsExample/index.js
@@ -36,5 +36,12 @@ export const SearchOptionsExample = ({ ordering }) => {
      */
     afterElements: [<p key>Test Element After</p>]
   }
-  return <SearchOptions {...props} />
+
+  const momentProps = {
+    minDate: new Date('2020-08-21 04:00 PM'),
+    maxDate: new Date('2020-08-27 04:00 PM'),
+    currentDate: new Date()
+  }
+
+  return <SearchOptions {...props} momentProps={momentProps} />
 }
